+++ db/migrate/20241121125140_change_default_value_of_bank_details_changing_task.rb
@@ -1,0 +1,5 @@
+class ChangeDefaultValueOfBankDetailsChangingTask < ActiveRecord::Migration[7.1]
+  def change
+    change_column :transfer_tasks_data, :bank_details_changing_yes_no, :boolean, default: nil
+  end
+end